refactor(platform/client): do not access host module from client module
