+++ projects/scion/microfrontend-platform/src/lib/host/message-broker/message-broker.ts
@@ -21,7 +21,8 @@ import {ClientRegistry} from '../client-registry/client.registry';
 import {chainInterceptors, IntentInterceptor, MessageInterceptor, PublishInterceptorChain} from './message-interception';
 import {Beans, Initializer, PreDestroy} from '@scion/toolkit/bean-manager';
 import {Runlevel} from '../../platform-state';
-import {APP_IDENTITY, ɵVERSION} from '../../platform.model';
+import {APP_IDENTITY} from '../../platform.model';
+import {ɵVERSION} from '../../ɵplatform.model';
 import {bufferUntil} from '@scion/toolkit/operators';
 import {filterByChannel, filterByTransport} from '../../operators';
 import {Client} from '../client-registry/client';
