+++ projects/scion/microfrontend-platform/src/lib/microfrontend-platform.ts
@@ -26,7 +26,8 @@ import {FocusInEventDispatcher} from './client/focus/focus-in-event-dispatcher';
 import {FocusMonitor} from './client/focus/focus-monitor';
 import {ContextService} from './client/context/context-service';
 import {RouterOutletUrlAssigner} from './client/router-outlet/router-outlet-url-assigner';
-import {APP_IDENTITY, IS_PLATFORM_HOST, ɵVERSION, ɵWINDOW_TOP} from './platform.model';
+import {APP_IDENTITY, IS_PLATFORM_HOST} from './platform.model';
+import {ɵVERSION, ɵWINDOW_TOP} from './ɵplatform.model';
 import {RelativePathResolver} from './client/router-outlet/relative-path-resolver';
 import {ClientRegistry} from './host/client-registry/client.registry';
 import {FocusTracker} from './host/focus/focus-tracker';
